Add Android App
This adds the current version of the Android app to the repository. This app is being built in Android Studio and uses Google Maps for the map view. Currently, this only shows hardcoded station information from the nrel.txt file found on the iOS app fork repository.

Note: This commit does not include the "/app/src/debug/res/values/google_maps_api.xml" file, which is required because the Google Maps API requires an API Key to work. This file is not provided because the API key should be kept private. You can message me later for help setting up this file on your devices.