Implement Basic Location Permissions Handling and Basic Dynamic Station List Querying
This adds in basic permissions handling for getting the user's current location. As of right now, this is not used yet beyond just showing the user where they are.

In addition, we now get the list of stations from the NREL database rather than hardcoding their values in. As of right now, this only gets the list of stations within 15 miles of the default location.

There is very limited error handling being done here too. This will need to be added later to ensure stability in nonideal situations.