Implement basic Bottom Sheets Window
Implements a basic bottom sheets window that is displayed whenever the user clicks on a marker. As of right now, the bottom sheets window does not display any meaningful information. To make up for this, current info windows are still shown. Once it is updated to display such information, it should look much sleeker and more modern than the current info windows.