Display more information in the bottom sheets view
This displays the remainder of the information we would like to see if the bottom sheets view when we select a station. This also fixes a bug with the charging level sorting based on preferences and also finishes implementation of sorting stations based on their charging availability. This also fixes the two buttons in the bottom sheets view.