Refactor Existing Source Code and Update Bottom Sheets
This refactors a lot of the code to either be more consistent in code style, to improve code readability, or to adhere to good programming practices. The appearance of the bottom sheets has also been updated and improved. The only new feature of this commit is the addition of a text description of where each station is physically located. More refactoring will be done later as some parts are still in need of it.