Continue Refactoring of Source Code
This continues refactoring the existing code to further adhere to good programming practices and also improves the ability to dynamically update the station markers on the map. This will be useful later for features such as updating the availability status of stations dynamically as it will allow us to do that without regenerating all the icons on the map.