Implement Notifications
Implements sliders on each station that the user can select / deselect to enable tracking for their charging or parking availability. Also implements a global option to track the availability of all nearby stations. When the status of these stations are updated, then the app sends a push notification to the user notifying them of this change.